requirements.txt
numpy>=1.26
pandas>=2.2
pyyaml>=6.0
scikit-learn>=1.5
tqdm>=4.66
sentence-transformers>=3.0
torch>=2.0
torchvision>=0.15
omegaconf>=2.3
timm>=0.9
diffusers>=0.25
lightning>=2.0
Pillow>=10.0
matplotlib>=3.8
gymnasium>=0.29
mani_skill>=3.0
